+++ lat.md/lat.md
@@ -8,3 +8,4 @@ This directory defines the high-level concepts, business logic, and architecture
 - [[window-chrome]] — the browser-style title bar where open-conversation tabs sit on top of the window drag region, clickable while empty space still drags.
 - [[sidebar-navigation]] — the recent-sessions list under the Chat nav item, capped at five with a "Show more" button that opens the full session list in a modal.
 - [[context-folder]] — the per-session linked working folder, persisted in a desktop-owned state.db table so a re-opened conversation restores its folder.
+- [[main-process]] — the Electron main-process entrypoint, app lifecycle modules, and centralized IPC registry.
