+++ lat.md/main-process.md
@@ -1,0 +1,27 @@
+# Main Process
+
+The Electron main process keeps the entrypoint small and separates app lifecycle from IPC registration.
+
+## Entrypoint
+
+`src/main/index.ts` performs only pre-ready setup and delegates startup.
+
+[[src/main/index.ts]] applies GPU crash preferences, enables the optional CDP testing port, and calls [[src/main/app/start.ts#startMainProcess]]. This keeps one-off process boot concerns separate from windows, menus, updater wiring, and IPC.
+
+## App Lifecycle
+
+Lifecycle code owns Electron windows, global app events, and shutdown cleanup.
+
+[[src/main/app/start.ts#startMainProcess]] registers crash logging, IPC handlers, updater handlers, Electron ready/activate/window-all-closed/before-quit events, CSP headers, security hardening, and the main BrowserWindow.
+
+## App Chrome Helpers
+
+Menu, updater, and context-menu behavior live in focused modules.
+
+[[src/main/app/menu.ts#buildMenu]] owns the application menu, [[src/main/app/updater.ts#setupUpdater]] owns update IPC and electron-updater events, and [[src/main/app/context-menu.ts#showChatContextMenu]] owns the chat right-click menu.
+
+## IPC Registry
+
+Renderer IPC handlers are isolated from app bootstrap so the registry can be split by domain.
+
+[[src/main/ipc/register.ts#registerIpcHandlers]] currently preserves the existing handler behavior behind one registration function. It receives app-level callbacks for the main window, model-library notifications, connection-config notifications, external URL opening, and active chat abort handles.
