+++ src/main/app/start.ts
@@ -1,0 +1,188 @@
+import {
+  app,
+  BrowserWindow,
+  session,
+  shell,
+} from "electron";
+import { join } from "path";
+import { electronApp, optimizer, is } from "@electron-toolkit/utils";
+import icon from "../../../resources/icon.png?asset";
+import { getPublicConnectionConfig } from "../config";
+import { stopHealthPolling } from "../hermes";
+import { stopAllDashboards } from "../dashboard";
+import { cleanupTempMediaFiles } from "../media";
+import { closeDbConnection } from "../db";
+import {
+  hardenAttachedWebContents,
+  hardenWebviewPreferences,
+  isAllowedAppNavigationUrl,
+  isAllowedExternalUrl,
+  isAllowedWebviewUrl,
+} from "../security";
+import { registerIpcHandlers } from "../ipc/register";
+import { showChatContextMenu } from "./context-menu";
+import { buildMenu } from "./menu";
+import { setupUpdater } from "./updater";
+
+const APP_NAME = process.env.HERMES_DESKTOP_APP_NAME?.trim() || "Hermes One";
+
+let mainWindow: BrowserWindow | null = null;
+const activeRuns = new Map<string, () => void>();
+
+export function startMainProcess(): void {
+  process.on("uncaughtException", (err) => {
+    console.error("[MAIN UNCAUGHT]", err);
+  });
+
+  process.on("unhandledRejection", (reason) => {
+    console.error("[MAIN UNHANDLED REJECTION]", reason);
+  });
+
+  registerIpcHandlers({
+    activeRuns,
+    getMainWindow: () => mainWindow,
+    notifyConnectionConfigChanged,
+    notifyModelLibraryChanged,
+    openExternalUrl,
+  });
+
+  setupUpdater({ getMainWindow: () => mainWindow });
+
+  app.whenReady().then(() => {
+    electronApp.setAppUserModelId("com.hermes.desktop");
+
+    app.on("browser-window-created", (_, window) => {
+      optimizer.watchWindowShortcuts(window);
+    });
+
+    app.on("web-contents-created", (_event, contents) => {
+      hardenAttachedWebContents(contents);
+    });
+
+    session.defaultSession.webRequest.onHeadersReceived((details, callback) => {
+      callback({
+        responseHeaders: {
+          ...details.responseHeaders,
+          "Content-Security-Policy": [
+            "default-src 'self'; " +
+              "script-src 'self' 'unsafe-inline'; " +
+              "style-src 'self' 'unsafe-inline'; " +
+              "img-src 'self' data: blob: file: https:; " +
+              "media-src 'self' data: blob: file: https:; " +
+              "connect-src 'self' http://127.0.0.1:* ws://127.0.0.1:* http://localhost:* ws://localhost:* https: wss:; " +
+              "font-src 'self' data:; " +
+              "frame-src 'self' https: http://127.0.0.1:* http://localhost:*; " +
+              "object-src 'none'; " +
+              "base-uri 'self';",
+          ],
+        },
+      });
+    });
+
+    createWindow();
+    buildMenu({ getMainWindow: () => mainWindow, openExternalUrl });
+
+    app.on("activate", () => {
+      if (BrowserWindow.getAllWindows().length === 0) createWindow();
+    });
+  });
+
+  app.on("window-all-closed", () => {
+    if (process.platform !== "darwin") app.quit();
+  });
+
+  app.on("before-quit", () => {
+    stopHealthPolling();
+    for (const abort of activeRuns.values()) abort();
+    activeRuns.clear();
+    cleanupTempMediaFiles();
+    stopAllDashboards();
+    closeDbConnection();
+  });
+}
+
+function notifyConnectionConfigChanged(): void {
+  mainWindow?.webContents.send(
+    "connection-config-changed",
+    getPublicConnectionConfig(),
+  );
+}
+
+function notifyModelLibraryChanged(): void {
+  mainWindow?.webContents.send("model-library-changed");
+}
+
+function openExternalUrl(rawUrl: unknown): void {
+  if (!isAllowedExternalUrl(rawUrl)) {
+    console.warn("[SECURITY] Blocked unsafe external URL");
+    return;
+  }
+  shell.openExternal(rawUrl).catch((err) => {
+    console.error("[SECURITY] Failed to open external URL:", err);
+  });
+}
+
+function createWindow(): void {
+  const rendererHtmlPath = join(__dirname, "../../renderer/index.html");
+  mainWindow = new BrowserWindow({
+    width: 1100,
+    height: 850,
+    minWidth: 900,
+    title: APP_NAME,
+    minHeight: 600,
+    show: false,
+    autoHideMenuBar: true,
+    titleBarStyle: process.platform === "darwin" ? "hiddenInset" : undefined,
+    ...(process.platform === "darwin"
+      ? { trafficLightPosition: { x: 16, y: 16 } }
+      : {}),
+    ...(process.platform === "linux" ? { icon } : {}),
+    webPreferences: {
+      preload: join(__dirname, "../../preload/index.js"),
+      nodeIntegration: false,
+      contextIsolation: true,
+      sandbox: true,
+      webSecurity: true,
+      allowRunningInsecureContent: false,
+      webviewTag: true,
+    },
+  });
+
+  mainWindow.on("ready-to-show", () => mainWindow?.show());
+  mainWindow.webContents.on("render-process-gone", (_event, details) => {
+    console.error("[CRASH] Renderer process gone:", details.reason, details.exitCode);
+  });
+  mainWindow.webContents.on("console-message", (_event, level, message, line, sourceId) => {
+    if (level >= 2) console.error(`[RENDERER ERROR] ${message} (${sourceId}:${line})`);
+  });
+  mainWindow.webContents.on("did-fail-load", (_event, errorCode, errorDescription) => {
+    console.error("[LOAD FAIL]", errorCode, errorDescription);
+  });
+  mainWindow.webContents.setWindowOpenHandler((details) => {
+    openExternalUrl(details.url);
+    return { action: "deny" };
+  });
+  mainWindow.webContents.on("will-navigate", (event, url) => {
+    if (isAllowedAppNavigationUrl(url, rendererHtmlPath, is.dev ? process.env["ELECTRON_RENDERER_URL"] : undefined)) return;
+    event.preventDefault();
+    openExternalUrl(url);
+  });
+  mainWindow.webContents.on("will-attach-webview", (event, webPreferences, params) => {
+    const isWebPreview = params.partition === "web-preview";
+    if (!isAllowedWebviewUrl(params.src, isWebPreview)) {
+      event.preventDefault();
+      console.warn("[SECURITY] Blocked webview attachment for untrusted URL");
+      return;
+    }
+    hardenWebviewPreferences(webPreferences);
+  });
+  mainWindow.webContents.on("context-menu", (_event, params) => {
+    showChatContextMenu(mainWindow, params);
+  });
+
+  if (is.dev && process.env["ELECTRON_RENDERER_URL"]) {
+    mainWindow.loadURL(process.env["ELECTRON_RENDERER_URL"]);
+  } else {
+    mainWindow.loadFile(rendererHtmlPath);
+  }
+}
