+++ src/main/app/updater.ts
@@ -1,0 +1,75 @@
+import { app, ipcMain, type BrowserWindow } from "electron";
+import type { AppUpdater } from "electron-updater";
+import { updaterLogger } from "../updater-log";
+
+interface UpdaterDeps {
+  getMainWindow: () => BrowserWindow | null;
+}
+
+export function setupUpdater({ getMainWindow }: UpdaterDeps): void {
+  ipcMain.handle("get-app-version", () => app.getVersion());
+
+  const isPortableBuild = !!process.env.PORTABLE_EXECUTABLE_DIR;
+  if (!app.isPackaged || isPortableBuild) {
+    ipcMain.handle("check-for-updates", async () => null);
+    ipcMain.handle("download-update", () => true);
+    ipcMain.handle("install-update", () => {});
+    return;
+  }
+
+  // eslint-disable-next-line @typescript-eslint/no-require-imports
+  const { autoUpdater } = require("electron-updater") as {
+    autoUpdater: AppUpdater;
+  };
+
+  autoUpdater.logger = updaterLogger;
+  autoUpdater.autoDownload = true;
+  autoUpdater.autoInstallOnAppQuit = true;
+
+  autoUpdater.on("update-available", (info) => {
+    getMainWindow()?.webContents.send("update-available", {
+      version: info.version,
+      releaseNotes: info.releaseNotes,
+    });
+  });
+  autoUpdater.on("download-progress", (progress) => {
+    getMainWindow()?.webContents.send("update-download-progress", {
+      percent: Math.round(progress.percent),
+    });
+  });
+  autoUpdater.on("update-downloaded", () => {
+    getMainWindow()?.webContents.send("update-downloaded");
+  });
+  autoUpdater.on("error", (err) => {
+    getMainWindow()?.webContents.send("update-error", err.message);
+  });
+
+  ipcMain.handle("check-for-updates", async () => {
+    try {
+      const result = await autoUpdater.checkForUpdates();
+      return result?.updateInfo?.version || null;
+    } catch {
+      return null;
+    }
+  });
+  ipcMain.handle("download-update", async () => {
+    try {
+      await autoUpdater.downloadUpdate();
+      return true;
+    } catch (err) {
+      const message = err instanceof Error ? err.message : String(err);
+      getMainWindow()?.webContents.send("update-error", message);
+      return false;
+    }
+  });
+  ipcMain.handle("install-update", () => {
+    updaterLogger.info(
+      "Restart requested by user — calling quitAndInstall(isSilent=false, isForceRunAfter=true)",
+    );
+    autoUpdater.quitAndInstall(false, true);
+  });
+
+  setTimeout(() => {
+    autoUpdater.checkForUpdates().catch(() => {});
+  }, 5000);
+}
