+++ src/renderer/src/screens/Chat/Chat.tsx
@@ -8,6 +8,7 @@ import { ModelPicker } from "./ModelPicker";
 import { ReasoningEffortPicker } from "./ReasoningEffortPicker";
 import { ContextFolderChip } from "./ContextFolderChip";
 import { WorktreePanel } from "./WorktreePanel";
+import { RemoteFolderPicker } from "./RemoteFolderPicker";
 import { WebPreviewPanel } from "./WebPreviewPanel";
 import { useChatScroll } from "./hooks/useChatScroll";
 import { useChatIPC } from "./hooks/useChatIPC";
@@ -161,6 +162,7 @@ function Chat({
   // Whether the worktree panel is visible (only applies when contextFolder is set)
   // Default false so the panel doesn't open automatically and interfere with scrolling
   const [worktreeVisible, setWorktreeVisible] = useState<boolean>(false);
+  const [folderPickerOpen, setFolderPickerOpen] = useState<boolean>(false);
   const [webPreviewVisible, setWebPreviewVisible] = useState<boolean>(false);
   const [webPreviewUrl, setWebPreviewUrl] =
     useState<string>("https://google.com");
@@ -638,9 +640,13 @@ function Chat({
   }, []);
 
   const handlePickFolder = useCallback(async () => {
+    if (remoteMode) {
+      setFolderPickerOpen(true);
+      return;
+    }
     const path = await window.hermesAPI.selectFolder();
     if (path) setContextFolder(path);
-  }, []);
+  }, [remoteMode]);
 
   const handleClearFolder = useCallback(() => {
     setContextFolder(null);
@@ -903,7 +909,7 @@ function Chat({
               </div>
               <ContextFolderChip
                 contextFolder={contextFolder}
-                show={!remoteMode}
+                show
                 worktreeVisible={worktreeVisible}
                 onPickFolder={handlePickFolder}
                 onClearFolder={handleClearFolder}
@@ -945,6 +951,15 @@ function Chat({
           </div>
         </div>
       )}
+      <RemoteFolderPicker
+        initialPath={contextFolder}
+        open={folderPickerOpen}
+        onCancel={() => setFolderPickerOpen(false)}
+        onSelect={(path) => {
+          setContextFolder(path);
+          setFolderPickerOpen(false);
+        }}
+      />
       {/* Show follow-us modal only after setup is complete */}
       {active && connectionModeLoaded && readiness.ok && <FollowUsModal />}
     </div>
